+++ src/Client/HttpClient.php
@@ -2,9 +2,11 @@
 
 namespace Akeneo\Pim\ApiClient\Client;
 
-use Http\Client\HttpClient as Client;
-use Http\Message\RequestFactory;
+use Psr\Http\Client\ClientInterface;
+use Psr\Http\Message\RequestFactoryInterface;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\StreamFactoryInterface;
+use Psr\Http\Message\StreamInterface;
 
 /**
  * Http client to send a request without any authentication.
@@ -15,25 +17,26 @@ use Psr\Http\Message\ResponseInterface;
  */
 class HttpClient implements HttpClientInterface
 {
-    /** @var Client */
+    /** @var ClientInterface */
     protected $httpClient;
 
-    /** @var RequestFactory */
+    /** @var RequestFactoryInterface */
     protected $requestFactory;
 
     /** @var HttpExceptionHandler */
     protected $httpExceptionHandler;
 
-    /**
-     * @param Client         $httpClient
-     * @param RequestFactory $requestFactory
-     */
+    /** @var StreamFactoryInterface */
+    private $streamFactory;
+
     public function __construct(
-        Client $httpClient,
-        RequestFactory $requestFactory
+        ClientInterface $httpClient,
+        RequestFactoryInterface $requestFactory,
+        StreamFactoryInterface $streamFactory
     ) {
         $this->httpClient = $httpClient;
         $this->requestFactory = $requestFactory;
+        $this->streamFactory = $streamFactory;
         $this->httpExceptionHandler = new HttpExceptionHandler();
     }
 
@@ -42,7 +45,19 @@ class HttpClient implements HttpClientInterface
      */
     public function sendRequest(string $httpMethod, $uri, array $headers = [], $body = null): ResponseInterface
     {
-        $request = $this->requestFactory->createRequest($httpMethod, $uri, $headers, $body);
+        $request = $this->requestFactory->createRequest($httpMethod, $uri);
+
+        if (null !== $body && is_string($body)) {
+            $request = $request->withBody($this->streamFactory->createStream($body));
+        }
+        if (null !== $body && $body instanceof StreamInterface) {
+            $request = $request->withBody($body);
+        }
+
+        foreach ($headers as $header => $content) {
+            $request = $request->withHeader($header, $content);
+        }
+
         $response = $this->httpClient->sendRequest($request);
         $response = $this->httpExceptionHandler->transformResponseToException($request, $response);